renamed post controller file
